# api-docs/swagger.json
{
  "swagger": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "api-docs"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
...
    "/users/registration": {
      "post": {
        "summary": "Create a new user",
        "operationId": "createUser",
        "tags": [
          "users"
        ],
        "security": [
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": "true",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "login user",
        "operationId": "loginUser",
        "tags": [
          "users"
        ],
        "security": [
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": "true",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    },
    "JWTAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "definitions": {
    "User": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "_id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "example": {
        "_id": "131231231",
        "username": "John",
        "email": "john@gmail.com",
        "password": "123456"
      }
    },
...
  }
}